Consolidate ajax error handling
