+++ resources/authgear/templates/en/web/__error.html
@@ -28,6 +28,7 @@
         <li id="error-message-server" class="hidden">{{ template "error-server" }}</li>
         <li id="error-message-network" class="hidden">{{ template "error-network" }}</li>
         <li id="error-message-invalid-selected-image" class="hidden">{{ template "error-invalid-selected-image" }}</li>
+        <li id="error-message-rate-limited" class="hidden">{{ template "error-rate-limited" }}</li>
         {{ if $display_error }}
             {{ $info := .Error.info }}
             {{ if eq .Error.reason "ValidationFailed" }}
